i3c_controller: include framing
Work in progress, this module joins the cmdp and sdio interface
into a single cmdw interface that inscructs the word (tbd) module
which "word" to send next to the bus.
At events when an ACK is not satisfied by a peripheral, the word
module will issue a _nack and the state-machine of the framing
will reset, cancelling the current word (will never Sr).

Signed-off-by: Jorge Marques <jorge.marques@analog.com>