axi_dac_interpolate: Add dma flush and raw feature
The flush feature can be used to clean the data from the DMA fifo.
This is usefull when the DMA is programed in cyclic mode and
data transmission is stoppet by dma_transfer_suspend flag
The software intervention is reduced to setting a flag.

Add raw value support. Through this changes a user can set
the dac output to a fixed predefined value in the following two cases:
1. direct, without using the dma.
2. with dma, as a hold value. The fixed value will be kipped after a cyclic
buffer is stopped by axi_dac_interpolate, through dma_transfer_suspend
register/signal.

The predefined value is stored in reg 0x19(0x64). For more details se
the documentation available at
https://wiki.analog.com/resources/fpga/docs/axi_dac_interpolate